+++ src/abstraction/eq4Transformateur1/Transformateur1Stocks.java
@@ -1,7 +1,6 @@
 package abstraction.eq4Transformateur1;
 
 import java.awt.Color;
-import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List; 
 
@@ -14,8 +13,10 @@ import abstraction.eqXRomu.general.Journal;
 import abstraction.eqXRomu.produits.Chocolat;
 import abstraction.eqXRomu.produits.ChocolatDeMarque;
 import abstraction.eqXRomu.produits.Feve;
+import abstraction.eqXRomu.produits.IProduit;
 import abstraction.eqXRomu.general.Variable;
 
+// Cette classe gère les stocks et controle les prix de vente
 public class Transformateur1Stocks extends Transformateur1Acteur implements IFabricantChocolatDeMarque {
 
 	//Des variables qui ne seront au final que des constantes lors de la simulation
@@ -23,136 +24,22 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 	private double coutProd;
 	protected double STOCK_MAX_TOTAL_FEVES = 1000000;
 
-	//Listes regroupant les contrats cadres actifs
-	protected List<ExemplaireContratCadre> mesContratEnTantQuAcheteur;
-	protected List<ExemplaireContratCadre> mesContratEnTantQueVendeur;
-
 	private List<ChocolatDeMarque> chocosProduits; // la liste de toutes les sortes de ChocolatDeMarque que l'acteur produit et peut vendre
-	protected HashMap<Feve, HashMap<Chocolat, Double>> pourcentageTransfo; // pour les differentes feves, le chocolat qu'elles peuvent contribuer a produire avec le ratio qttChocoProduit/qttFevesUtilisée
-
-	//Des tables de hachages pour connaître l'état des chocolats à une période précise
-	protected HashMap<Feve, Double> qttEntrantesFeve;//Contient les quantités entrant dans le stock à la période actuelle
-	protected HashMap<Feve, Double> prixTFeveStockee;//Contient les prix moyens des fèves en stock
-	protected HashMap<Chocolat, Double> prixTChocoBase;//Contient les prix des chocolats produits en s'appuyant sur le prix du stock de fèves
-	protected HashMap<Chocolat, Double> qttSortantesChoco;
-	protected HashMap<Chocolat, Double> marges;
-	protected Variable qttFevesAcheteesBourse;
-	protected HashMap<Chocolat, Double> qttSortantesTransactions;
-
-	protected Variable prix_Limdt_MQ;
-	protected Variable prix_Limdt_BQ_E; 
-	protected Variable prix_Limdt_MQ_E;
-	protected Variable prix_Limdt_HQ_BE;
-
 
 	public Transformateur1Stocks() {
 		super();
-
 		this.chocosProduits = new LinkedList<ChocolatDeMarque>();
-		this.pourcentageTransfo = new HashMap<Feve, HashMap<Chocolat, Double>>();
-		
-
-		this.qttEntrantesFeve = new HashMap<Feve, Double>();
-		this.prixTFeveStockee = new HashMap<Feve, Double>();
-		this.prixTChocoBase = new HashMap<Chocolat, Double>();
-		this.qttFevesAcheteesBourse = new Variable("Qtt Feves Achetees Bourse", "<html>Quantité de fèves achetées en bourse</html>", this, 0., 1000000., 1000.);
-
-		this.prix_Limdt_MQ = new Variable("Prix LimDt MQ", "<html>Prix de vente du chocolat de marque MQ</html>", this, 0., 1000000., 0.);
-		this.prix_Limdt_BQ_E = new Variable("Prix LimDt BQ_E", "<html>Prix de vente du chocolat de marque BQ_E</html>", this, 0., 1000000., 0.);
-		this.prix_Limdt_MQ_E = new Variable("Prix LimDt MQ_E", "<html>Prix de vente du chocolat de marque MQ_E</html>", this, 0., 1000000., 0.);
-		this.prix_Limdt_HQ_BE = new Variable("Prix LimDt HQ_BE", "<html>Prix de vente du chocolat de marque HQ_BE</html>", this, 0., 1000000., 0.);
-
-		this.qttSortantesChoco = new HashMap<Chocolat, Double>();
-		this.qttSortantesTransactions = new HashMap<Chocolat, Double>();
-
-		this.marges = new HashMap<Chocolat, Double>();
-
 	}
 
-
-
-
-
-	
 	public void initialiser() {
 		super.initialiser();
-
-		//Initialisation des prix de nos stocks de fève
-		this.prixTFeveStockee.put(Feve.F_MQ, 2000.);
-		this.prixTFeveStockee.put(Feve.F_BQ_E, 2000.);
-		this.prixTFeveStockee.put(Feve.F_MQ_E, 2000.);
-		this.prixTFeveStockee.put(Feve.F_HQ_BE, 2000.);
-
 		this.coutStockage = Filiere.LA_FILIERE.getParametre("cout moyen stockage producteur").getValeur()*4;
 		this.coutProd = 4000; //A MODIFIER il s'agit du cout de la production d'une tonne de chocolat, valeur arbitraire censée contenir salaires, ingrédients secondaires, et autres couts fixes
-
-		//Initialisation des prix de base des chocolats que l'on veut produire
-		this.prixTChocoBase.put(Chocolat.C_MQ, 2000.);
-		this.prix_Limdt_MQ.setValeur(this, 2000.);
-
-		this.prixTChocoBase.put(Chocolat.C_BQ_E, 2000.);
-		this.prix_Limdt_BQ_E.setValeur(this, 2000.);
-
-		this.prixTChocoBase.put(Chocolat.C_MQ_E, 2000.);
-		this.prix_Limdt_MQ_E.setValeur(this, 2000.);
-
-		this.prixTChocoBase.put(Chocolat.C_HQ_BE, 2000.);
-		this.prix_Limdt_HQ_BE.setValeur(this, 2000.);
-		
-
-		//Initialisation des marges que l'on va faire sur les différents produits
-		this.marges.put(Chocolat.C_MQ, 1.5);
-		this.marges.put(Chocolat.C_BQ_E, 1.16);
-		this.marges.put(Chocolat.C_MQ_E, 1.16);
-		this.marges.put(Chocolat.C_HQ_BE, 1.3);
-
-
-		//Initialisation des pourcentage de conversion fèves vers chocolat
-		this.pourcentageTransfo.put(Feve.F_HQ_BE, new HashMap<Chocolat, Double>());
-		double conversion = 1.0 + (100.0 - Filiere.LA_FILIERE.getParametre("pourcentage min cacao HQ").getValeur())/100.0;
-		this.pourcentageTransfo.get(Feve.F_HQ_BE).put(Chocolat.C_HQ_BE, conversion);// la masse de chocolat obtenue est plus importante que la masse de feve vue l'ajout d'autres ingredients
-
-		this.pourcentageTransfo.put(Feve.F_MQ_E, new HashMap<Chocolat, Double>());
-		conversion = 1.0 + (100.0 - Filiere.LA_FILIERE.getParametre("pourcentage min cacao MQ").getValeur())/100.0;
-		this.pourcentageTransfo.get(Feve.F_MQ_E).put(Chocolat.C_MQ_E, conversion);
-
-		this.pourcentageTransfo.put(Feve.F_MQ, new HashMap<Chocolat, Double>());
-		this.pourcentageTransfo.get(Feve.F_MQ).put(Chocolat.C_MQ, conversion);
-
-		this.pourcentageTransfo.put(Feve.F_BQ_E, new HashMap<Chocolat, Double>());
-		conversion = 1.0 + (100.0 - Filiere.LA_FILIERE.getParametre("pourcentage min cacao BQ").getValeur())/100.0;
-		this.pourcentageTransfo.get(Feve.F_BQ_E).put(Chocolat.C_BQ_E, conversion);
-
-
-		this.journalStock.ajouter(Romu.COLOR_LLGRAY, Color.PINK, "Stock initial chocolat de marque : ");
-
-		this.journalCC.ajouter(Color.orange, Color.BLACK, "Les achats seront en marron;");
-		this.journalCC.ajouter(Color.orange, Color.BLACK, "Les ventes LimDt en mauve;");
-		this.journalCC.ajouter(Color.orange, Color.BLACK, "Et les autres ventes en vert.");
-		this.journalCC.ajouter("\n");
-
-		this.journalTransactions.ajouter(Color.orange, Color.BLACK, "Les achats en bourse seront en magenta;");
-		this.journalTransactions.ajouter(Color.orange, Color.BLACK, "Les ventes aux enchères en gris foncé;");
-		this.journalTransactions.ajouter(Color.orange, Color.BLACK, "Et les ventes AO en rouge.");
-		this.journalTransactions.ajouter("\n");
-
-		//Initialisation des quantités de fève entrantes
-		this.qttEntrantesFeve.put(Feve.F_MQ, 0.);
-		this.qttEntrantesFeve.put(Feve.F_BQ_E, 0.);
-		this.qttEntrantesFeve.put(Feve.F_HQ_BE, 0.);
-		this.qttEntrantesFeve.put(Feve.F_MQ_E, 0.);
 	}
-
-
-
-
 
 	////////////////////////////////////////////////////////
 	//      En lien avec la comptabilité et production    //
 	////////////////////////////////////////////////////////
-
-
-
 
 	/**
 	 * @author ABBASSI Rayene
@@ -174,13 +61,9 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 					//On s'assure que l'on produit quelque chose pour faire nos opérations
 					if (transfo<=this.getQuantiteEnStock(f, this.cryptogramme) && transfo >0) {
 
-
-
 						double pourcentageMarque = 1.0;  //Modifiable
 						// La Pourcentage ainsi definie sera stockee sous forme de marquee, la quantité restante sera alors stockee comme non marquee
 
-	
-						
 						//A MODIFIER
 						int pourcentageCacao =  (int) (Filiere.LA_FILIERE.getParametre("pourcentage min cacao "+c.getGamme()).getValeur());
 						ChocolatDeMarque cm= new ChocolatDeMarque(c, "LimDt", pourcentageCacao);
@@ -211,7 +94,6 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 						//this.ajouterAuStock(c, nouveauStock * (1.0-pourcentageMarque), this.cryptogramme);
 						this.ajouterAuStock(cm, nouveauStock * pourcentageMarque, this.cryptogramme);
 
-						
 						//Notification dans le journal
 						this.journal.ajouter(Romu.COLOR_LLGRAY, Color.PINK, "Transfo de "+(transfo<10?" "+transfo:transfo)+" T de "+f+" en "+Journal.doubleSur(transfo*this.pourcentageTransfo.get(f).get(c),3,2)+" T de "+c);
 						this.journal.ajouter(Romu.COLOR_LLGRAY, Color.BLACK," stock("+f+")->"+this.getQuantiteEnStock(f, this.cryptogramme));
@@ -224,15 +106,6 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 		}
 	}
 
-
-
-
-
-
-
-
-
-
 	/**
 	 * @author MURY Julien
 	 * @author YAOU Reda
@@ -242,9 +115,6 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 	 * @return None
 	 */
 	public void determinerQttEntrantFeves(){
-		for (Feve f : pourcentageTransfo.keySet()){
-			this.qttEntrantesFeve.put(f, 0.);
-		}
 
 		//Réinitialisation des qttEntrantes
 		for(Feve f : this.lesFeves){
@@ -272,12 +142,7 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 		} else {
 			this.qttEntrantesFeve.put(Feve.F_MQ, qttFevesAcheteesBourse.getValeur());
 		}
-
-
-
 	}
-
-
 
 	/**
 	 * @author MURY Julien
@@ -288,33 +153,28 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 	 */
 	public double determinerQttEntrantFevesAuStep(int step, Feve f){
 
-
 		double qttEntrante = 0.;
 
 		//Quantité entrante de fèves f par contrat cadre
-			
 		for (ExemplaireContratCadre cc : mesContratEnTantQuAcheteur){
 			if (cc.getProduit() == f){
 				qttEntrante+= cc.getEcheancier().getQuantite(step);
 			}
 		}
 	
-
 		//Quantité entrante de fèves par achat en bourse (On approxime par le fait que la quantité de feve achetée en bourse est constante pour les autres step)
 		qttEntrante += qttFevesAcheteesBourse.getValeur();
 
 		return qttEntrante;
-
 	}
 
-
-
-
-/**
- * @author MURY Julien
- */
+   /**
+    * @author MURY Julien
+	* @author YAOU Reda
+    */
 	public void determinerQttSortantChoco(){
 
+		//Réinitialisation des qttSortantes
 		for (Chocolat c : lesChocolats){
 			this.qttSortantesChoco.put(c, 0.);
 		}
@@ -333,6 +193,7 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 			}
 		}
 
+		//Qtt sortante par vente en transaction (enchéres + AO)
 		for (Chocolat c : lesChocolats){
 			if (this.qttSortantesChoco.containsKey(c)){
 				this.qttSortantesChoco.put(c, this.qttSortantesTransactions.get(c) +this.qttSortantesChoco.get(c));
@@ -340,13 +201,8 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 			else {
 				this.qttSortantesChoco.put(c, this.qttSortantesTransactions.get(c));
 			}
-			
 		}
 	}
-
-
-
-
 
 	/**
 	 * @author MURY Julien
@@ -359,13 +215,8 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 				qttSortant += cc.getEcheancier().getQuantite(step);
 			}
 		}
-
 		return qttSortant;
-		
 	}
-
-
-
 
 
 	/**
@@ -404,24 +255,19 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 				
 				prixTFeveStockee.put(f, ancienPrixPondere + prix*(qttEntrantesFeve.get(f)/(qttEntrantesFeve.get(f)+getQuantiteEnStock(f, this.cryptogramme))));
 			}
-			
 		}
 	}
-
-
-
-
-
-
-
-
-
-
 
 	////////////////////////////////////////////////////////
 	//         En lien avec l'interface graphique         //
 	////////////////////////////////////////////////////////
 
+	/**
+	 * @author MURY Julien
+	 * @author ABASSI Rayene
+	 * @author YAOU Reda : Gestion des journaux
+	 * Cette méthode est appelée à chaque étape de la simulation. Elle permet de faire avancer le temps et de mettre à jour les stocks et les prix.
+	 */
 	public void next() {
 		super.next();
 
@@ -440,24 +286,22 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 			this.qttSortantesTransactions.put(c, 0.);
 		}
 
+		this.journal.ajouter("##########");
 		this.journal.ajouter(Color.yellow, Romu.COLOR_LBLUE, "N° Etape " + Filiere.LA_FILIERE.getEtape());
 
 		this.journalStock.ajouter("\n");
+		this.journalStock.ajouter("##########");
 		this.journalStock.ajouter(Color.yellow, Romu.COLOR_LBLUE, "N° Etape " + Filiere.LA_FILIERE.getEtape());
 
+		this.journalCC.ajouter("##########");
 		this.journalCC.ajouter(Color.yellow, Romu.COLOR_LBLUE, "N° Etape " + Filiere.LA_FILIERE.getEtape());
 
+		this.journalTransactions.ajouter("##########");
 		this.journalTransactions.ajouter(Color.yellow, Romu.COLOR_LBLUE, "N° Etape " + Filiere.LA_FILIERE.getEtape());
 
 		this.journalPeremption.ajouter("\n");
+		this.journalPeremption.ajouter("##########");
 		this.journalPeremption.ajouter(Color.yellow, Romu.COLOR_LBLUE, "N° Etape " + Filiere.LA_FILIERE.getEtape());
-
-		//OBSOLETE ou A MODIFIER
-		/*
-		this.journalStock.ajouter("Stock de fèves : " + this.totalStocksFeves.getValeur(this.cryptogramme));
-		this.journalStock.ajouter("Stock de chocolat : " + this.totalStocksChoco.getValeur(this.cryptogramme));
-		this.journalStock.ajouter("\n");
-		*/
 
 		//Affichage des stocks de chaque produit dans le journalStock à la période présente 
 		this.journal.ajouter("=== STOCKS === ");
@@ -476,6 +320,7 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 		}
 		this.journalStock.ajouter("\n");
 
+		// Affichage de l'état de péremption des stocks de chocolat de marque
 		this.journalPeremption.ajouter(Romu.COLOR_LLGRAY, Color.BLACK, "Péremption C_MQ_Limdt : ");
 		for (int i=0; i<12; i++) {
 			this.journalPeremption.ajouter(Romu.COLOR_LLGRAY, Color.BLACK, i+" : "+this.péremption_C_MQ_Limdt[i]);
@@ -500,6 +345,7 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 		}
 		this.journalPeremption.ajouter("\n");
 
+		// Détermination de prix des fèves et transformation
 		this.determinerPrixTFevesStockees();
 		this.transformation();
 
@@ -556,6 +402,7 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 					break;
 			}
 		}
+
 		//Calcul des stocks globaux pour payer le cout du stockage
 		double totalStocks = 0;
 		for (Feve f : this.lesFeves){
@@ -568,16 +415,9 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 			totalStocks += this.getQuantiteEnStock(cm, this.cryptogramme);
 		}
 
-
+        //On paye le cout de stockage
 		Filiere.LA_FILIERE.getBanque().payerCout(this, cryptogramme, "Stockage", (totalStocks*this.coutStockage));
-
-		//System.out.println("Voici nos prix : " + prixTChocoBase);
 	}
-
-
-
-
-
 
 
 	public List<Variable> getIndicateurs(){
@@ -604,12 +444,9 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 		return res;
 	}
 
-
-
 	////////////////////////////////////////////////////////
 	//        Pour la creation de filieres de test        //
 	////////////////////////////////////////////////////////
-
 
 	public List<String> getMarquesChocolat() {
 		LinkedList<String> marques = new LinkedList<String>();
@@ -627,4 +464,198 @@ public class Transformateur1Stocks extends Transformateur1Acteur implements IFab
 		return this.chocosProduits;
 	}
 
+    /**
+     * @author MURY Julien
+	 * @author YAOU Reda : Gestion de la péremption
+     * Cette méthode permet l'ajout en stock d'une certaine quantité d'un certain produit
+     * @param produit : le produit à ajouter
+     * @param quantite : la quantité de produit à ajouter 
+     */
+	public void ajouterAuStock(IProduit produit, double quantite, int cryptogramme){
+		//On vérifie que l'acteur voulant accéder aux stocks est bien le bon 
+		if (this.cryptogramme == cryptogramme){
+
+			String type = produit.getType();
+
+			if (type.equals("Feve")){
+				switch ((Feve)produit){
+					case F_MQ : 
+						this.stocksFevesVar.get(Feve.F_MQ).ajouter(this, quantite, cryptogramme);
+						break;
+					case F_BQ_E : 
+						this.stocksFevesVar.get(Feve.F_BQ_E).ajouter(this, quantite, cryptogramme);
+						break;
+					case F_MQ_E : 
+						this.stocksFevesVar.get(Feve.F_MQ_E).ajouter(this, quantite, cryptogramme);
+						break;
+					case F_HQ_BE : 
+						this.stocksFevesVar.get(Feve.F_HQ_BE).ajouter(this, quantite, cryptogramme);
+						break;
+					default : 
+						journalStock.ajouter(Color.pink, Color.BLACK, "EQ4T : Ce type de fève n'est pas censée entrer dans nos stocks: " + produit);
+				}
+				
+			}
+			else if (type.equals("Chocolat")){
+				switch ((Chocolat)produit){
+					case C_MQ : 
+						this.stocksChocoVar.get(Chocolat.C_MQ).ajouter(this, quantite, cryptogramme);
+						break;
+					case C_BQ_E : 
+						this.stocksChocoVar.get(Chocolat.C_BQ_E).ajouter(this, quantite, cryptogramme);
+						break;
+					case C_MQ_E : 
+						this.stocksChocoVar.get(Chocolat.C_MQ_E).ajouter(this, quantite, cryptogramme);
+						break;
+					case C_HQ_BE : 
+						this.stocksChocoVar.get(Chocolat.C_HQ_BE).ajouter(this, quantite, cryptogramme);
+						break;
+					default : 
+						journalStock.ajouter(Color.pink, Color.BLACK,"EQ4T : Ce type de chocolat n'est pas censée entrer dans nos stocks: " + produit);
+				}
+
+			}
+			else if (type.equals("ChocolatDeMarque")){
+				switch (((ChocolatDeMarque)produit).getChocolat()){
+					case C_MQ : 
+						this.stocksMarqueVar.get(produit).ajouter(this, quantite, cryptogramme);
+						this.péremption_C_MQ_Limdt[0] += quantite;
+						break;
+					case C_BQ_E : 
+						this.stocksMarqueVar.get(produit).ajouter(this, quantite, cryptogramme);
+						this.péremption_C_BQ_E_Limdt[0] += quantite;
+						break;
+					case C_MQ_E : 
+						this.stocksMarqueVar.get(produit).ajouter(this, quantite, cryptogramme);
+						this.péremption_C_MQ_E_Limdt[0] += quantite;
+						break;
+					case C_HQ_BE : 
+						this.stocksMarqueVar.get(produit).ajouter(this, quantite, cryptogramme);
+						this.péremption_C_HQ_BE_Limdt[0] += quantite;
+						break;
+					default : 
+						journalStock.ajouter(Color.pink, Color.BLACK,"EQ4T : Ce type de chocolat n'est pas censée entrer dans nos stocks: " + produit);
+				}
+			}
+			else {
+				journalStock.ajouter(Color.pink, Color.BLACK,"Ce produit (" + produit + ") n'a pas un type connu ("+ type + ")");
+			}
+		}
+	}
+
+    /**
+     * @author MURY Julien
+	 * @author YAOU Reda : Gestion de la péremption
+     * Cette méthode permet le retrait du stock d'une certaine quantité d'un certain produit
+     * @param produit : le produit à ajouter
+     * @param quantite : la quantité de produit à ajouter 
+     */
+	public void retirerDuStock(IProduit produit, double quantite, int cryptogramme){
+		//On vérifie que l'acteur voulant accéder aux stocks est bien le bon 
+		if (this.cryptogramme == cryptogramme){
+			String type = produit.getType();
+			if (type.equals("Feve")){
+				switch ((Feve)produit){
+					case F_MQ : 
+						this.stocksFevesVar.get(Feve.F_MQ).retirer(this, quantite, cryptogramme);
+						break;
+					case F_BQ_E : 
+						this.stocksFevesVar.get(Feve.F_BQ_E).retirer(this, quantite, cryptogramme);
+						break;
+					case F_MQ_E : 
+						this.stocksFevesVar.get(Feve.F_MQ_E).retirer(this, quantite, cryptogramme);
+						break;
+					case F_HQ_BE : 
+						this.stocksFevesVar.get(Feve.F_HQ_BE).retirer(this, quantite, cryptogramme);
+						break;
+					default : 
+						journalStock.ajouter(Color.pink, Color.BLACK,"EQ4T : Ce type de fève n'est pas censée entrer dans nos stocks: " + produit);
+				}
+				
+			}
+			else if (type.equals("Chocolat")){
+				switch ((Chocolat)produit){
+					case C_MQ : 
+						this.stocksChocoVar.get(Chocolat.C_MQ).retirer(this, quantite, cryptogramme);
+						break;
+					case C_BQ_E : 
+						this.stocksChocoVar.get(Chocolat.C_BQ_E).retirer(this, quantite, cryptogramme);
+						break;
+					case C_MQ_E : 
+						this.stocksChocoVar.get(Chocolat.C_MQ_E).retirer(this, quantite, cryptogramme);
+						break;
+					case C_HQ_BE : 
+						this.stocksChocoVar.get(Chocolat.C_HQ_BE).retirer(this, quantite, cryptogramme);
+						break;
+					default : 
+						journalStock.ajouter(Color.pink, Color.BLACK,"EQ4T : Ce type de chocolat n'est pas censée entrer dans nos stocks: " + produit);
+				}
+	
+			}
+			else if (type.equals("ChocolatDeMarque")){
+				switch (((ChocolatDeMarque)produit).getChocolat()){
+					case C_MQ : 
+						this.stocksMarqueVar.get(produit).retirer(this, quantite, cryptogramme);
+						for (int i=11; i>=0; i--){
+							if (this.péremption_C_MQ_Limdt[i] > 0 && this.péremption_C_MQ_Limdt[i] - quantite >= 0){
+								this.péremption_C_MQ_Limdt[i] -= quantite;
+								break;
+							}
+							else if (this.péremption_C_MQ_Limdt[i] > 0 && this.péremption_C_MQ_Limdt[i] - quantite < 0){
+								quantite -= this.péremption_C_MQ_Limdt[i];
+								this.péremption_C_MQ_Limdt[i] = 0;
+							}
+						}
+						break;
+					case C_BQ_E : 
+						this.stocksMarqueVar.get(produit).retirer(this, quantite, cryptogramme);
+						for (int i=11; i>=0; i--){
+							if (this.péremption_C_BQ_E_Limdt[i] > 0 && this.péremption_C_BQ_E_Limdt[i] - quantite >= 0){
+								this.péremption_C_BQ_E_Limdt[i] -= quantite;
+								break;
+							}
+							else if (this.péremption_C_BQ_E_Limdt[i] > 0 && this.péremption_C_BQ_E_Limdt[i] - quantite < 0){
+								quantite -= this.péremption_C_BQ_E_Limdt[i];
+								this.péremption_C_BQ_E_Limdt[i] = 0;
+							}
+						}
+						break;
+					case C_MQ_E : 
+						this.stocksMarqueVar.get(produit).retirer(this, quantite, cryptogramme);
+						for (int i=11; i>=0; i--){
+							if (this.péremption_C_MQ_E_Limdt[i] > 0 && this.péremption_C_MQ_E_Limdt[i] - quantite >= 0){
+								this.péremption_C_MQ_E_Limdt[i] -= quantite;
+								break;
+							}
+							else if (this.péremption_C_MQ_E_Limdt[i] > 0 && this.péremption_C_MQ_E_Limdt[i] - quantite < 0){
+								quantite -= this.péremption_C_MQ_E_Limdt[i];
+								this.péremption_C_MQ_E_Limdt[i] = 0;
+							}
+						}
+						break;
+					case C_HQ_BE : 
+						this.stocksMarqueVar.get(produit).retirer(this, quantite, cryptogramme);
+						for (int i=11; i>=0; i--){
+							if (this.péremption_C_HQ_BE_Limdt[i] > 0 && this.péremption_C_HQ_BE_Limdt[i] - quantite >= 0){
+								this.péremption_C_HQ_BE_Limdt[i] -= quantite;
+								break;
+							}
+							else if (this.péremption_C_HQ_BE_Limdt[i] > 0 && this.péremption_C_HQ_BE_Limdt[i] - quantite < 0){
+								quantite -= this.péremption_C_HQ_BE_Limdt[i];
+								this.péremption_C_HQ_BE_Limdt[i] = 0;
+							}
+						}
+						break;
+					default : 
+						journalStock.ajouter(Color.pink, Color.BLACK,"EQ4T : Ce type de chocolat n'est pas censée entrer dans nos stocks: " + produit);
+				}
+			}
+			else {
+				journalStock.ajouter(Color.pink, Color.BLACK,"Ce produit (" + produit + ") n'a pas un type connu ("+ type + ")");
+			}
+		}
+	}
 }
+
+
+
