+++ src/abstraction/eq4Transformateur1/Transformateur1VendeurAppelDoffre.java
@@ -16,21 +16,19 @@ import abstraction.eqXRomu.produits.ChocolatDeMarque;
 
 public class Transformateur1VendeurAppelDoffre extends Transformateur1AcheteurBourse implements IVendeurAO {
 
-
     public Transformateur1VendeurAppelDoffre() {
 		super();
 	}
-
 
 	public void initialiser(){
 		super.initialiser();
 	}
 
-
 	@Override
 	public OffreVente proposerVente(AppelDOffre offre) {
-		//System.err.println(offre.toString());
+;
 		double prixT = 0;
+
 		if (chocolatsLimDt.contains(offre.getProduit())) {
 			if (offre.getQuantiteT() <= 0.4*this.getQuantiteEnStock(offre.getProduit(), this.cryptogramme)) {
 				if (((ChocolatDeMarque) offre.getProduit()).getChocolat() == Chocolat.C_MQ) {
@@ -62,14 +60,12 @@ public class Transformateur1VendeurAppelDoffre extends Transformateur1AcheteurBo
 
 	@Override
 	public void notifierVenteAO(OffreVente propositionRetenue) {
-		//System.out.println("Votre proposition de vente a été retenue");
+
 		this.journalTransactions.ajouter(Romu.COLOR_LLGRAY, Color.RED, "-->AO "+ propositionRetenue.getProduit() +" : J'ai vendu " +propositionRetenue.getQuantiteT()+ " tonnes au prix de " + propositionRetenue.getPrixT() + " euros par tonne.");
 		this.journalTransactions.ajouter("\n");
 	 
-		//Mettre à jour les autres variables
+		//Mettre à jour le stock et la qtt sortante par transaction
 		ChocolatDeMarque chocoMarqueAO = (ChocolatDeMarque) propositionRetenue.getProduit();
-
-
 		this.retirerDuStock(chocoMarqueAO, propositionRetenue.getQuantiteT(), this.cryptogramme);
 		this.qttSortantesTransactions.put(chocoMarqueAO.getChocolat(), this.qttSortantesTransactions.get(chocoMarqueAO.getChocolat()) + propositionRetenue.getQuantiteT());
 	}
@@ -77,18 +73,10 @@ public class Transformateur1VendeurAppelDoffre extends Transformateur1AcheteurBo
 
 	@Override
 	public void notifierPropositionNonRetenueAO(OffreVente propositionRefusee) {
-		//System.out.println("Votre proposition de vente n'a pas été retenue");
+
 		this.journalTransactions.ajouter(Color.pink, Color.RED, "--> AO "+propositionRefusee.getProduit()+" : Vente non retenue.");
 		this.journalTransactions.ajouter("\n");
-		/*if (((ChocolatDeMarque) propositionRefusee.getProduit()).getChocolat() == Chocolat.C_MQ) {
-			prix_MQ = prix_MQ*0.95;
-		} else if (((ChocolatDeMarque) propositionRefusee.getProduit()).getChocolat() == Chocolat.C_BQ_E) {
-			prix_BQ_E -= 50;
-		} else if (((ChocolatDeMarque) propositionRefusee.getProduit()).getChocolat() == Chocolat.C_MQ_E) {
-			prix_MQ_E -= 50;
-		} else if (((ChocolatDeMarque) propositionRefusee.getProduit()).getChocolat() == Chocolat.C_HQ_BE) {
-			prix_HQ_BE -= 50;
-		}*/
+
 	}
 }
